requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23,<1.0
pydantic>=2.5,<3
pydantic-settings>=2.2,<3
python-multipart>=0.0.9,<0.0.10
httpx>=0.26,<0.28
SQLAlchemy>=2.0,<3
sqlmodel>=0.0.16,<0.0.22
charset-normalizer>=3.0,<4
python-docx>=0.8.11,<1.1
mammoth>=1.6,<2
docx2python>=2.8,<3
pdfplumber>=0.11,<0.12
pdfminer.six>=20221105
PyMuPDF>=1.23,<1.25
pikepdf>=8,<10
pytest>=8,<9
reportlab>=3.6,<4
